Move diversity.py to visualisation
